Adding and deleting functionality for Ionic App
Uploaded screenshots for ionic app. Added functionality of adding and deleting items in an array. Alert tells user that the item is deleted. When user clicks on the "plus" icon on the bottom right an alert pops up and inquires name and quantity of new item and gives option to either add the item to the array or to cancel interaction. Used interpolation through ionic to reduce repeatable code.